requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic>=2.9.0
pydantic-settings>=2.1.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
bcrypt==4.1.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
requests==2.31.0
openpyxl==3.1.2
xlrd==2.0.1

# ML stack
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.4.2
joblib==1.4.2
pyarrow==16.1.0
geoalchemy2==0.14.3
